PCM-1509 fixed default audioStream to not tear down after session ends. refactored test-utils and mock-apis
