+++ doc/index.md
@@ -227,6 +227,11 @@ defaults?: {
 A default audio stream to accept softphone sessions with
   if no audio stream was used when accepting the session
   (ie: `sdk.acceptSession({ id: 'session-id', mediaStream })`)
+
+> Warning: Firefox does not allow multiple microphone media tracks. 
+using a default could cause the SDK to be unable to request any 
+other audio device besides the active microphone – which would be the 
+audio track on this default stream. 
           
 #### `defaults.audioElement`
 `audioElement?: HTMLAudioElement;` Optional: no default. (See note about default behavior)
