+++ src/types/interfaces.ts
@@ -150,6 +150,11 @@ export interface ISdkConfig {
      *  if no audio stream was used when accepting the session
      *  (ie: `sdk.acceptSession({ id: 'session-id', mediaStream })`)
      *
+     * Warning: Firefox does not allow multiple microphone media tracks.
+     *  using a default could cause the SDK to be unable to request any
+     *  other audio device besides the active microphone - which would be the
+     *  audio track on this default stream.
+     *
      * Optional: no default
      */
     audioStream?: MediaStream;
