+++ test/test-pages/index-template.html
@@ -79,6 +79,8 @@
       </form>
     </div>
     <button id="start-app-button" type="button" class="btn btn-success">Start App</button>
+    <button id="start-app-button-with-default-audio" type="button" class="btn btn-info">Start App w/ Default
+      Audio</button>
 
     <div class="acd-screen-share">
       <div>
